pyproject.toml
[tool.poetry]
name = "firebot-tours"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
python-dotenv = "^0.15.0"
beeprint = "^2.4.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
